Dockerfile: Remove pthreadpool-git from full variant
This is causing build failures recently due to an issue in googletest that this package tries to download. We required this as an dependency for mlgo training but since mlgo updated their tflite dependencies that got rid of pthreadpool lib requirement, we no longer need this library. Remove it and solve build issues.